Convert duration function into scanDuration
Move the duration function into reports scaninfo module because it is
only used at this place. Also it doesn't display a formatted duration
instead it calculates a duration between two dates and converts it to a
readable string more or less specific for scan durations.
